[WIP] Add basic frame extractor script
This one doesn't really do anything yet, except being able to parse
incoming arguments and having chosen a pcap parsing library.

`pcapkit` will be used for now, as it is able to support the newer
pcapng file format whilst having a broad range of parsing engines (like
dpkt - which is the fastest according to them). A better solution would
be a fast streaming parser to reduce memory footprint, but the parsing
is fast enoughish that the current process will be fine for now.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>