[WIP] Working (but useless) RV908 simulator
This adds an overall working RV908 simulator that contains very
primitive packet definitions (for now), but can render incoming frames
in real-time, and evaluate and verify packets (must be extended).

As of now there's no state machine nor the ability to send data
integrated, but this will come in the future. As it is mostly a
proof-of-concept for one of the harder engineering parts (processing the
overwhelming amount of packets and rendering them in real time), there
is still a lot missing.

Behind the scenes:
- it doesn't use the wireshark dissector engine anymore as it was far to
  slow for our purposes (and writing lua definitions with data verifications
  was a pain). So we switched to scapy.
- the GUI is Tkinter with an OpenGL widget which renders the frames and
  runs in a separate process. The GIL made multi-threading not a
  feasible solution (I tried)
- The packet definition now contains test methods that allow us to write
  atomized and simple verifications for the incoming data. This is
  mostly in order to find the errors in our own assumptions and correct
  the definitions / tests accordingly.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>