[WIP] improve simulator
= 1. Add state machine for 980 board
It is currently unclear whether this is actually necessary. There's a
slow suspicious creeping that the protocol is entirely stateless and
as such no state machine is necessary. But it needs to be mentioned that
not everything of the protocol has been explored so far. So there might
be stateful behaviors somewhere in it (e.g. updating, …)

To keep everything organized, the simulator was put into its own module.

= 2. Add receiver packet parser/builder
For scapy there's no difference between defining how a packet will be
parsed and how it will be built. So a basic receiver packet definition
was added to send our own packets when necessary (doesn't seem to be,
but this can change over time).

= 3. Heavily improve the sender packet parser
During exploration of the protocol there were a few new protocol
features discovered. Mainly:
- there are no broadcast messages per se
- there's a complex command structure integrated in the header, with
  data, checksum, …

We were also able to RE the checksum of the protocol that is used in the
header. Yay!

= 4. Enabled destructive scapy parsing errors
Normally the dissector fail silently - probably because in most cases
it's not the dissectors fault, but the users. But in our case the data
is the ground truth not our shitty representation of how it might work.
So we enabled raising dissector errors globally. (see main.py)

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>