[WIP] simulator: Add memory representation
With this we can now finally fully interpret the memory segments that
are transferred through the cmds data packets (in the header) during
setting of things.

To make it far easier to raffle through the sheer amount of data that
the memory consists of, a custom memory / pattern format was introduced
(see memory.hex). This one allows us to define static parts of the
memory that shouldn't change and known parameters with their associated
parsers.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>