Simulator: Multi-line patterns and gamma curves
This commit mostly adds parsing of single channel gamma curves from
the (simulated) RV908 memory. In order to do this, the hex pattern
language needed to be extended to allow multi-line patterns. An example
can be seen in `memory.hex`, where there are four pattern descriptors
defined that span multiple lines (one for each of the four channels).

It also became obvious that the RV908 internally uses four color
channels (whatever the fourth one is). As a lot of parameters can be
individually configured per channel, we introduce a data structure that
makes it much easier to create such a parameter (represented by python).

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>