[fix] sim: Prevent OpenGL start for test execution
When the package `pyopengltk` is imported, it will automatically try to
start an OpenGL context in the background, even if it is never used.
This fails, when OpenGL does not work rn due to various circumstances.

To prevent this we will only (indirectly) import the above package, when
the UI is actually initialized, allowing us to test the code or work
headless without worrying about OpenGL incompatibilities.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>