[feat] sim: Improve image frame filtering
For performance reasons, we can't just parse all incoming frames using
Scapy (as has been discussed previously). The solution we found was to
not parse any irrelevant frames and only extract the image part from
those. With this commit we refine that filtering by adding a rule to
interpret the config data pkg index so that we can skip interpreting
the repeating config data, which would otherwise clog up our pipeline.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>