[feat] sim: Correct CmdsConfigData definition
Analysis of the config packages turned up, that the previous definitions
of CmdsConfigData were incorrect / incomplete. For one: The delimiter is
bollocks, it is the highest byte of the three byte address - instead of
some random stuff.

Also, the flag is presumably just an enum. This makes the parsing /
verifying and such much easier in the long run. Interpreting it as
flags, did not make sense with the introduction of volatile safes.

Using this the verification of the CmdsConfigData has been overhauled
completely and tested.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>